renaming omnimath to omni_math
