+++ src/helm/benchmark/annotation/omni_math/gpt_evaluation_template.txt
