+++ src/helm/benchmark/annotation/omni_math_annotator.py
@@ -22,13 +22,13 @@ def parse_report(report):
 
 
 class OmniMATHAnnotator(Annotator):
-    """The OmniMATH autograder."""
+    """The Omni-MATH autograder."""
 
-    name = "omnimath"
+    name = "omni_math"
 
     def __init__(self, auto_client: AutoClient):
         self._auto_client = auto_client
-        template_path = files("src.helm.benchmark.annotation.omnimath").joinpath("gpt_evaluation_template.txt")
+        template_path = files("src.helm.benchmark.annotation.omni_math").joinpath("gpt_evaluation_template.txt")
         with template_path.open("r") as file:
             self._score_template = file.read()
