+++ src/helm/benchmark/run_specs/lite_run_specs.py
@@ -513,24 +513,24 @@ def get_bigcodebench_spec(version: str) -> RunSpec:
     )
 
 
-@run_spec_function("omnimath")
-def get_omnimath_spec() -> RunSpec:
+@run_spec_function("omni_math")
+def get_omni_math_spec() -> RunSpec:
 
     scenario_spec = ScenarioSpec(
-        class_name="helm.benchmark.scenarios.omnimath_scenario.OmniMATHScenario"
+        class_name="helm.benchmark.scenarios.omni_math_scenario.OmniMATHScenario"
     )
 
     adapter_spec = AdapterSpec(
         method=ADAPT_GENERATION, input_prefix="", output_prefix="", max_tokens=1000, num_outputs=1, temperature=0.0,
     )
-    annotator_specs = [AnnotatorSpec(class_name="helm.benchmark.annotation.omnimath_annotator.OmniMATHAnnotator")]
-    metric_specs = get_basic_metric_specs([]) + [MetricSpec(class_name="helm.benchmark.metrics.omnimath_metrics.OmniMATHMetric")]
+    annotator_specs = [AnnotatorSpec(class_name="helm.benchmark.annotation.omni_math_annotator.OmniMATHAnnotator")]
+    metric_specs = get_basic_metric_specs([]) + [MetricSpec(class_name="helm.benchmark.metrics.omni_math_metrics.OmniMATHMetric")]
 
     return RunSpec(
-        name="omnimath",
+        name="omni_math",
         scenario_spec=scenario_spec,
         adapter_spec=adapter_spec,
         annotators=annotator_specs,
         metric_specs=metric_specs,
-        groups=["omnimath"],
+        groups=["omni_math"],
     )
